ensurefloat redesigned
Former-commit-id: c58cfe3caa526dfd274d9f02f31b7cdff30c7530

+++ src/lib/Tools.h
@@ -30,7 +30,7 @@ namespace inscore
 struct Tools
 {
 	static bool			regexp (const std::string& str);
-	static std::string	ensurefloat (double f, int precision=10);
+	static std::string	ensurefloat (float f, int precision=0);
 };
